Refactor validation scripts to synchronize host-call counts
- Updated validation scripts (validate_r2303 to validate_r2314) to read host calls from the correct source files and check against the package host call count instead of the runtime required host call count.
- Removed unnecessary runtime required host call count variable from scripts.
- Adjusted assertions in validation functions to ensure consistency with the new host call checks.
- Enhanced CLI help documentation to include experimental feature options and commands.
- Improved code readability and performance in various sections of the CLI and formatter modules.

+++ packages/spectra-api/src/trace.rs
@@ -10,7 +10,7 @@ unsafe fn string(value: i64) -> Option<String> {
     let ptr = value as *const u8;
     let mut bytes = Vec::new();
     for index in 0..4096 {
-        let byte = *ptr.add(index) as u8;
+        let byte = *ptr.add(index);
         if byte == 0 {
             return String::from_utf8(bytes).ok();
         }
